--- a/goby.go
+++ b/goby.go
@@ -11,6 +11,7 @@ import (
 	"github.com/goby-lang/goby/compiler"
 	"github.com/goby-lang/goby/compiler/parser"
 	"github.com/goby-lang/goby/igb"
+	_ "github.com/goby-lang/goby/lib"
 	"github.com/goby-lang/goby/vm"
 	"github.com/pkg/profile"
 )