--- a/native/db/db_test.go
+++ b/native/db/db_test.go
@@ -1,4 +1,4 @@
-package lib
+package db
 
 import (
 	"testing"