--- a/vm/block.go
+++ b/vm/block.go
@@ -2,6 +2,7 @@ package vm
 
 import (
 	"fmt"
+
 	"github.com/goby-lang/goby/vm/classes"
 	"github.com/goby-lang/goby/vm/errors"
 )
@@ -19,9 +20,9 @@ func builtinBlockClassMethods() []*BuiltinMethodObject {
 		{
 			Name: "new",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 					if blockFrame == nil {
-						return t.vm.initErrorObject(errors.ArgumentError, sourceLine, "Can't initialize block object without block argument")
+						return t.vm.InitErrorObject(errors.ArgumentError, sourceLine, "Can't initialize block object without block argument")
 					}
 
 					return t.vm.initBlockObject(blockFrame.instructionSet, blockFrame.ep, blockFrame.self)
@@ -37,7 +38,7 @@ func builtinBlockInstanceMethods() []*BuiltinMethodObject {
 		{
 			Name: "call",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 					block := receiver.(*BlockObject)
 					c := newNormalCallFrame(block.instructionSet, block.instructionSet.filename, sourceLine)
 					c.ep = block.ep
@@ -84,7 +85,7 @@ func (bo *BlockObject) toString() string {
 }
 
 // toJSON just delegates to toString
-func (bo *BlockObject) toJSON(t *thread) string {
+func (bo *BlockObject) toJSON(t *Thread) string {
 	return bo.toString()
 }