--- a/vm/boolean.go
+++ b/vm/boolean.go
@@ -27,7 +27,7 @@ func builtinBooleanClassMethods() []*BuiltinMethodObject {
 		{
 			Name: "new",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 					return t.vm.initUnsupportedMethodError(sourceLine, "#new", receiver)
 				}
 			},
@@ -48,7 +48,7 @@ func builtinBooleanInstanceMethods() []*BuiltinMethodObject {
 			// @return [Boolean]
 			Name: "==",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 
 					if receiver == args[0] {
 						return TRUE
@@ -68,7 +68,7 @@ func builtinBooleanInstanceMethods() []*BuiltinMethodObject {
 			// @return [Boolean]
 			Name: "!=",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 
 					if receiver != args[0] {
 						return TRUE
@@ -87,7 +87,7 @@ func builtinBooleanInstanceMethods() []*BuiltinMethodObject {
 			// @return [Boolean]
 			Name: "!",
 			Fn: func(receiver Object, sourceLine int) builtinMethodBody {
-				return func(t *thread, args []Object, blockFrame *normalCallFrame) Object {
+				return func(t *Thread, args []Object, blockFrame *normalCallFrame) Object {
 
 					rightValue := receiver.(*BooleanObject).value
 
@@ -139,7 +139,7 @@ func (b *BooleanObject) toString() string {
 }
 
 // toJSON just delegates to `toString`
-func (b *BooleanObject) toJSON(t *thread) string {
+func (b *BooleanObject) toJSON(t *Thread) string {
 	return b.toString()
 }