--- a/vm/diggable.go
+++ b/vm/diggable.go
@@ -2,5 +2,5 @@ package vm
 
 // Diggable represents a class that support the #dig method.
 type Diggable interface {
-	dig(t *thread, keys []Object, sourceLine int) Object
+	dig(t *Thread, keys []Object, sourceLine int) Object
 }