--- a/vm/error.go
+++ b/vm/error.go
@@ -3,8 +3,9 @@ package vm
 import (
 	"fmt"
 
-	"github.com/goby-lang/goby/vm/errors"
 	"strings"
+
+	"github.com/goby-lang/goby/vm/errors"
 )
 
 // Error class is actually a special struct to hold internal error types with messages.
@@ -29,10 +30,10 @@ type Error struct {
 // Functions for initialization -----------------------------------------
 
 func (vm *VM) initUnsupportedMethodError(sourceLine int, methodName string, receiver Object) *Error {
-	return vm.initErrorObject(errors.UnsupportedMethodError, sourceLine, "Unsupported Method %s for %+v", methodName, receiver.toString())
+	return vm.InitErrorObject(errors.UnsupportedMethodError, sourceLine, "Unsupported Method %s for %+v", methodName, receiver.toString())
 }
 
-func (vm *VM) initErrorObject(errorType string, sourceLine int, format string, args ...interface{}) *Error {
+func (vm *VM) InitErrorObject(errorType string, sourceLine int, format string, args ...interface{}) *Error {
 	errClass := vm.objectClass.getClassConstant(errorType)
 
 	t := vm.mainThread
@@ -73,7 +74,7 @@ func (e *Error) toString() string {
 }
 
 // toJSON just delegates to `toString`
-func (e *Error) toJSON(t *thread) string {
+func (e *Error) toJSON(t *Thread) string {
 	return e.toString()
 }