This is a very large diff, and most of the changes are exporting internal functions so other packages can use them. In my opinion, the surface area of the API should be smaller, but for now I think exporting these functions is fine. There are also some general code quality changes as well. 

Main changes:
* Functions for registering native libraries.
* Testing helpers for external libraries.

Perf related changes include:
* VM's main thread is no longer a pointer to a thread.
* Threads no longer need any initialization logic. `var t Thread` only needs it's vm to be set before it is functional
* Thread's stack and call stack are no longer pointers but live inside of the thread itself.

```bash
➜  goby git:(thread-rework) ./benchmark.rb
benchmarking b6a1996c7832449e3c1bb71c61d6b867c91901cc
benchmarking da753973fd6100c9fc4317a9adb5c42da0d8d3d9
universal.x86_64-darwin17
benchmark                         old ns/op     new ns/op     delta
BenchmarkBasicMath/add-8          5959          5424          -8.98%
BenchmarkBasicMath/subtract-8     5918          5379          -9.11%
BenchmarkBasicMath/multiply-8     6075          5521          -9.12%
BenchmarkBasicMath/divide-8       5829          5448          -6.54%

benchmark                         old allocs     new allocs     delta
BenchmarkBasicMath/add-8          78             78             +0.00%
BenchmarkBasicMath/subtract-8     78             78             +0.00%
BenchmarkBasicMath/multiply-8     78             78             +0.00%
BenchmarkBasicMath/divide-8       78             78             +0.00%

benchmark                         old bytes     new bytes     delta
BenchmarkBasicMath/add-8          3792          3792          +0.00%
BenchmarkBasicMath/subtract-8     3792          3792          +0.00%
BenchmarkBasicMath/multiply-8     3792          3792          +0.00%
BenchmarkBasicMath/divide-8       3792          3792          +0.00%
Switched to branch 'thread-rework'
```