@st0012 2018-05-05
@eliothedeman does the performance gain come from replacing pointer with object itself?